Split nerves.ex into multiple files
