+++ lib_src/nerves_runtime/exit.ex
@@ -1,0 +1,9 @@
+defmodule Toolshed.NervesRuntime.Exit do
+  @doc """
+  Exit the current IEx session
+  """
+  @spec exit() :: true
+  def exit() do
+    Process.exit(Process.group_leader(), :kill)
+  end
+end
